+++ FitnessApp/Home/Models/Workout.swift
@@ -1,0 +1,19 @@
+//
+//  Workout.swift
+//  FitnessApp
+//
+//  Created by APPLE on 30/07/24.
+//
+
+import Foundation
+import SwiftUI
+
+struct Workout{
+    let id:Int
+    let title:String
+    let image:String
+    let tintColor:Color
+    let duration:String
+    let date:String
+    let calories:String
+}
