+++ FitnessApp/Home/View/HomeView.swift
@@ -7,26 +7,10 @@
 
 import SwiftUI
 
+
 struct HomeView: View {
-    @State var calories: Int = 123
-    @State var active: Int = 123
-    @State var stand: Int = 123
-    
-    var mockActivities = [
-        Activity(id: 0, title: "Today steps", subtitle: "Goal 12,000", image: "figure.walk", tintColor: .green, amount: "9,812"),
-        Activity(id: 1, title: "Today steps", subtitle: "Goal 12,000", image: "figure.walk", tintColor: .red, amount: "9,812"),
-        Activity(id: 2, title: "Today steps", subtitle: "Goal 12,000", image: "figure.walk", tintColor: .blue, amount: "9,812"),
-        Activity(id: 3, title: "Today steps", subtitle: "Goal 12,000", image: "figure.walk", tintColor: .purple, amount: "9,812")
-    ]
-    
-   var mockWorkouts = [
-    Workout(id: 0, title: "Running", image: "figure.run", tintColor: .cyan, duration: "51 mins", date: "Aug 1", calories: "512 kcal"),
-    Workout(id: 1, title: "Strength", image: "figure.run", tintColor: .red, duration: "51 mins", date: "Aug 1", calories: "512 kcal"),
-    Workout(id: 2, title: "Cycling", image: "figure.run", tintColor: .blue, duration: "51 mins", date: "Aug 1", calories: "512 kcal"),
-    Workout(id: 3, title: "Yoga", image: "figure.run", tintColor: .purple, duration: "51 mins", date: "Aug 1", calories: "512 kcal"),
-  
-   ]
-    var body: some View {
+    @State var viewModel = HomeViewModel()
+   var body: some View {
         NavigationStack {
             ScrollView(showsIndicators: false){
                 VStack{
@@ -36,14 +20,14 @@ struct HomeView: View {
                     
                     HStack{
                         Spacer()
-                        VStack{
+                        VStack(alignment: .leading){
                             VStack(alignment: .leading,spacing: 8){
                                 Text("Calories")
                                     .font(.callout)
                                     .bold()
                                     .foregroundColor(.red)
                                 
-                                Text("123 kcal")
+                                Text("\(viewModel.calories)")
                                     .bold()
                             }
                             .padding(.bottom)
@@ -54,7 +38,7 @@ struct HomeView: View {
                                     .bold()
                                     .foregroundColor(.green)
                                 
-                                Text("52 mins")
+                                Text("\(viewModel.active)")
                                     .bold()
                             }
                             .padding(.bottom)
@@ -65,18 +49,18 @@ struct HomeView: View {
                                     .bold()
                                     .foregroundColor(.blue)
                                 
-                                Text("8 hours")
+                                Text("\(viewModel.stand)")
                                     .bold()
                             }
                         }
                         Spacer()
                         
                         ZStack{
-                            ProgressCircleView(progress: $calories, goal: 600, color: .red)
+                            ProgressCircleView(progress: $viewModel.calories, goal: 600, color: .red)
                             
-                            ProgressCircleView(progress: $calories, goal: 60, color: .green)
+                            ProgressCircleView(progress: $viewModel.calories, goal: 60, color: .green)
                                 .padding(.all,20)
-                            ProgressCircleView(progress: $calories, goal: 12, color: .blue)
+                            ProgressCircleView(progress: $viewModel.calories, goal: 12, color: .blue)
                                 .padding(.all,40)
                         }
                         .padding(.horizontal)
@@ -102,7 +86,7 @@ struct HomeView: View {
                     .padding(.horizontal)
                     
                     LazyVGrid(columns:Array(repeating: GridItem(spacing:20), count: 2)){
-                        ForEach(mockActivities, id:\.id){activity in
+                        ForEach(viewModel.mockActivities, id:\.id){activity in
                             ActivityCard(activity: activity)
                         }
                     }
@@ -130,7 +114,7 @@ struct HomeView: View {
                 .padding(.top)
                  
                 LazyVStack{
-                    ForEach(mockWorkouts,id: \.id){ workout in
+                    ForEach(viewModel.mockWorkouts,id: \.id){ workout in
                         WorkoutCard(workout: workout)
                     }
                 }
